--- a/include/mapnik/timer.hpp
+++ b/include/mapnik/timer.hpp
@@ -69,7 +69,7 @@ struct timer_metrics {
 class timer_stats_
 {
 public:
-    void add(std::string metric_name, double cpu_elapsed, double wall_clock_elapsed)
+    void add(std::string const& metric_name, double cpu_elapsed, double wall_clock_elapsed)
     {
         timer_metrics& metrics = timer_stats_[metric_name];
         metrics.cpu_elapsed += cpu_elapsed;
@@ -77,7 +77,7 @@ public:
         timer_stats_[metric_name] = metrics;
     }
 
-    timer_metrics get(std::string metric_name)
+    timer_metrics get(std::string const& metric_name)
     {
         return timer_stats_[metric_name];
     }