Hey, I just wanted you to take a quick look at these ideas in code. The code is crappy, lacks tests, breaks for sure (as I modified the API of the stats) and can be improved.

I tested it using a file like this:
```c++
#include <unistd.h>
#include <iostream>
#include "include/mapnik/timer.hpp"

/**
 * Compile it like this:
 * g++-4.9 -std=c++11 metrics-test.cpp -o metrics-test
 */
int main(int argc, char** argv)
{
    {
        mapnik::progress_timer __stats__("dummy::sleep");
        usleep(100000);
    }

    //mapnik::timer_stats.reset("dummy::sleep");

    {
        mapnik::progress_timer __stats__("dummy::sleep");
        usleep(100000);
    }

    //mapnik::timer_stats.reset_all();

    {
        mapnik::progress_timer __stats__("busy::wait");
        unsigned long int counter = 0;
        for(unsigned int i=0; i<10000; i++) {
            for(unsigned int j=0; j<10000; j++) {
                counter++;
            }
        }
    }

    for(auto metric : mapnik::timer_stats) {
        std::cout << metric.first << "\tcpu_time = " << metric.second.cpu_elapsed << " ms\twall_time = " << metric.second.wall_clock_elapsed << " ms" << std::endl;
    }


    // std::cout << "dummy::sleep.cpu_elapsed = " << mapnik::timer_stats.get("dummy::sleep").cpu_elapsed << " ms" << std::endl;
    // std::cout << "dummy::sleep.wall_clock_elapsed = " << mapnik::timer_stats.get("dummy::sleep").wall_clock_elapsed << " ms" << std::endl;
    // std::cout << std::endl;
    // std::cout << "busy::wait.cpu_elapsed = " << mapnik::timer_stats.get("busy::wait").cpu_elapsed << " ms" << std::endl;
    // std::cout << "busy::wait.wall_clock_elapsed = " << mapnik::timer_stats.get("busy::wait").wall_clock_elapsed << " ms" << std::endl;

    return 0;
}
```

and running it on my machine outputs this stuff:
```
$ time ./metrics-test 
busy::wait	cpu_time = 379.447 ms	wall_time = 379.572 ms
dummy::sleep	cpu_time = 0.052 ms	wall_time = 200.225 ms

real	0m0.583s
user	0m0.380s
sys	0m0.000s
```

The idea is to use the same code to collect stats and expose the `get` or some other custom method to spit them out to wrappers and upper layers.

Maybe instead of just an unordered_map (hash) we'd need to use an LRU in-memory cache or something like that (circular buffer, whatever, you name it) to avoid big memory footprint. Maybe it is better to have a separate class to gather these stats in production and be more selective.

Comments in this early stage are more than welcome cc/ @javisantana @jgoizueta